requirements.txt
django>=4.2.0,<5.0.0
djangorestframework
pyyaml
django-allauth
djangorestframework-simplejwt
celery
django-cors-headers
pytest
psycopg2-binary
requests